+++ StoreApp/StoreApp/Model/StoreItems.swift
@@ -10,9 +10,34 @@ import Foundation
 
 class StoreItems {
     
-    var bests = [StoreItem]()
-    var masks = [StoreItem]()
-    var grocerys = [StoreItem]()
-    var fryingpans = [StoreItem]()
+    private(set) var bests = [StoreItem]()
+    private(set) var masks = [StoreItem]()
+    private(set) var grocerys = [StoreItem]()
+    private(set) var fryingpans = [StoreItem]()
+    
+    func addData(category: String, items: [StoreItem]) {
+        
+        switch category {
+        case "best":
+            for item in items {
+                self.bests.append(item)
+            }
+        case "mask":
+            for item in items {
+                self.masks.append(item)
+            }
+        case "grocery":
+            for item in items {
+                self.grocerys.append(item)
+            }
+        case "fryingpan":
+            for item in items {
+                self.fryingpans.append(item)
+            }
+        default:
+            return
+        }
+        
+    }
     
 }
