+++ StoreApp/StoreApp/View/Network/NetworkManager.swift
@@ -15,15 +15,9 @@ class NetworkManager {
         let url = "https://public.codesquad.kr/jk/kakao-2021/\(category).json"
         URLSession(configuration: URLSessionConfiguration.default).dataTask(with: URL(string: url)!) {
             (data, response, error) in
-            guard let json = data else { return }
-            let decoder = JSONDecoder()
+            guard let jsonData = data else {return}
             DispatchQueue.main.async {
-                do {
-                    let results = try decoder.decode([StoreItem].self, from: json)
-                    NotificationCenter.default.post(name: Notification.Name(rawValue: "getItemData"), object: JsonData(category, results), userInfo: nil)
-                } catch {
-                    print(error.localizedDescription)
-                }
+                NotificationCenter.default.post(name: Notification.Name(rawValue: "getItemData"), object: jsonData, userInfo: ["category":category])
             }
         }.resume()
